pretty sick.  So i guess theres a problem with xcache_unset_by_prefix().  maybe its not in 1.2.2 but i think theres problems with it in 1.3.0 so i made a work around where everything is stored in an array so if you happen to update a post when theres a bunch of cache (when the gc hasn't come along and cleaned stuff up) it will unset each one.

+++ private/includes/filecache.php
@@ -7,7 +7,7 @@
  */
 class filecache
 {
-	private $cacheloc = null;
+	private $cacheLoc = null;
 	private $uri = null;
 	private $urimd5 = null;
 	private $fileloc = null;
@@ -24,7 +24,7 @@ class filecache
 	public function __construct($uri)
 	{
 		// setting up the variables
-		$this->cacheloc = V_BASELOC . "/private/cache/";
+		$this->cacheLoc = V_BASELOC . "/private/cache/";
 		//echo $this->cacheloc;
 		$this->uri = $uri;
 		// magic quotes?
@@ -32,7 +32,8 @@ class filecache
 		if(get_magic_quotes_gpc())
 			$this->uri = stripcslashes($this->uri);
 		$this->urimd5 = md5($this->uri);
-		$this->fileloc = $this->cacheloc . "/" . $this->urimd5;
+		$this->fileloc = $this->cacheLoc . "/" . $this->urimd5;
+
 	}
 	
 	
@@ -174,7 +175,7 @@ class filecache
 		
 		while($file = readdir($dir))
 		{
-			if($file != "." && $file != "..")
+			if($file != "." && $file != ".." && $file != ".svn")
 			{
 				unlink($this->cacheLoc . "/" . $file);
 			}
